Обнова
Крупное обновление, другая структура, больше компонентов
