+++ src/js/View/Components/Section 1/MainPage.jsx
