+++ src/js/View/Components/Section 2/Cites.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+
+class Cites extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div className="cites">
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            <a href="https://torrent-games.co/" target="_blank"><img src="img/cite.png" alt="cite"/></a>
+            </div>
+        );
+    }
+}
+
+export default Cites;
