+++ src/js/View/Components/Section 4/Contact.jsx
@@ -11,12 +11,12 @@ class Contact extends React.Component{
                 <div className="contact-wrap wrap">
                     <h1 className="contact-title">CONTACT US</h1>
                     <h2 className="contact-desc">Oh how lovely that you want to talk with me, because I want to talk with you too! XOXO</h2>
-                        <form className="callback" >
-                            <textarea className="name" type="text" placeholder="YOUR NAME"></textarea>
-                            <textarea className="mail" type="text" placeholder="YOUR MAIL"></textarea>
-                            <textarea className="message" type="text" placeholder="YOUR MESSAGE"></textarea>
+                        <form action="php/mail.php" method="post" className="callback" >
+                            <textarea className="name" type="text" name="name" placeholder="YOUR NAME"></textarea>
+                            <textarea className="mail" type="text" name="email" placeholder="YOUR MAIL"></textarea>
+                            <textarea className="message" type="text" name="message" placeholder="YOUR MESSAGE"></textarea>
                         </form>
-                    <button className="send-message">SEND MESSAGE</button>
+                    <button type="submit" className="send-message">SEND MESSAGE</button>
                 </div>
             </div>
         );
